+++ compiler/src/mir/passes/mod.rs
@@ -101,8 +101,8 @@ pub(crate) fn find_min_max_addresses<'m>(
             match repeat.source {
                 RepeatSource::Count(count) => {
                     let count_0_address = total_address_offsets + address.value;
-                    let count_max_address =
-                        count_0_address + (i128::from(count.saturating_sub(1)) * repeat.stride);
+                    let count_max_address = count_0_address
+                        + (i128::from(count.saturating_sub(1)) * repeat.stride as i128);
                     let min_address = count_0_address.min(count_max_address);
                     let max_address = count_0_address.max(count_max_address);
 
@@ -123,8 +123,9 @@ pub(crate) fn find_min_max_addresses<'m>(
                         .expect("A mir pass checked this is an enum");
 
                     for (discriminant, _) in enum_value.iter_variants_with_discriminant() {
-                        let address =
-                            total_address_offsets + address.value + (discriminant * repeat.stride);
+                        let address = total_address_offsets
+                            + address.value
+                            + (discriminant * repeat.stride) as i128;
                         if address < min_address_found {
                             min_address_found = address;
                             min_obj_found = Some(object);
